# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Investment Alphas testing API",
    "description": "API for testing trading alphas",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/forward/start": {
      "post": {
        "tags": [
          "forward"
        ],
        "summary": "Start Forward Test",
        "description": "Initialize a new forward test",
        "operationId": "start_forward_test_api_v1_forward_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForwardTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "forward:write"
            ]
          }
        ]
      }
    },
    "/api/v1/forward/stop": {
      "put": {
        "tags": [
          "forward"
        ],
        "summary": "Stop Forward Test",
        "description": "Stop forward testing for an account",
        "operationId": "stop_forward_test_api_v1_forward_stop_put",
        "security": [
          {
            "OAuth2PasswordBearer": [
              "forward:write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID of the account to stop forward testing",
              "title": "Account Id"
            },
            "description": "ID of the account to stop forward testing"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forward/history/{account_id}": {
      "get": {
        "tags": [
          "forward"
        ],
        "summary": "Get Forward Test History",
        "description": "Get portfolio value history for a forward test",
        "operationId": "get_forward_test_history_api_v1_forward_history__account_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": [
              "forward:read"
            ]
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forward/active": {
      "get": {
        "tags": [
          "forward"
        ],
        "summary": "List Active Forward Tests",
        "description": "List all active forward tests for the current user",
        "operationId": "list_active_forward_tests_api_v1_forward_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "forward:read"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ForwardTestRequest": {
        "properties": {
          "instruments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Instruments",
            "description": "List of instrument tickers to use in forward test"
          },
          "alpha_id": {
            "type": "integer",
            "title": "Alpha Id",
            "description": "ID of the alpha to use in forward test"
          },
          "trade_on_weekends": {
            "type": "boolean",
            "title": "Trade On Weekends",
            "description": "Whether to trade on weekends",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "instruments",
          "alpha_id"
        ],
        "title": "ForwardTestRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {
              "users:read": "Read information about users",
              "users:write": "Create, update, and delete users",
              "alphas:read": "Read information about trading alphas",
              "alphas:write": "Create, update, and delete trading alphas",
              "backtest:read": "View backtest results",
              "backtest:write": "Run backtests",
              "forward:read": "View forward test results",
              "forward:write": "Run forward tests"
            },
            "tokenUrl": "api/v1/auth/token"
          }
        }
      }
    }
  }
}